requirements.txt
Flask>=2.0
Pillow>=9.0
pikepdf>=9.0